git-svn-id: https://pascalinterpreterinjava.googlecode.com/svn/trunk@90 fb5b4802-194b-0410-9e11-5134f6b347ec

+++ edu/js/interpreter/plugins/scar_conversions.java
@@ -11,7 +11,7 @@ public class scar_conversions implements pascal_plugin {
 		return String.valueOf(i);
 	}
 
-	public static String FloatToStr(float f) {
+	public static String FloatToStr(double f) {
 		return String.valueOf(f);
 	}
